requirements.txt
# Web frameworks
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0

# Environment variables
python-dotenv==1.0.1

#LangGraph
langgraph>=0.2.0
langgraph>=0.2.0

# LLM clients
openai==1.58.0
mistralai==1.5.0

# GitHub API
PyGithub==2.4.0

# HTTP requests (backend + frontend)
requests==2.32.3

# Pydantic for models / validation
pydantic==2.9.0  # Updated to match the required range

# CLI helpers
click==8.1.7

setuptools<82.0.0
wheel